DS3232: Change the maths for the temperature calculation
The way I did it before might have given the wrong result for temps between -1 and 0, because we would've lost the sign bit on the initial
conversion of the integer part to a signed integer, and then added the fractional part instead of subtracting it. So -0.75 would have been
returned as 0.75. The new way is tidier anyway.