Fix implementation of getTemperature() so negative values work.
The old implementation would fail for negative values because it directly cast the integer part from an unsigned 8-bit int into a float (doing
entirely the wrong thing for a negative two's complement value), and then added the fractional part.
A value of -0.75C would be reported as 255.25C!

The new implementation handles both positive and negative values correctly.