Stop doing pointless (and wrong) conversion of day-of-the-week
This is just a number from 0 to 6 as far as the RTCs are concerned, and they just increment it at midnight, wrapping back round from 6 to 0.
Applications are free to assign whatever day of the week to those numbers they want. Converting from 1 to 7 was never correct.
Converting to/from BCD was also pointless as it's always a single digit.